rename planning_validator collision_checker module
