--- planning/planning_validator/autoware_planning_validator_collision_checker/plugins.xml
@@ -1,3 +1,0 @@
-<library path="autoware_planning_validator_collision_checker">
-  <class type="autoware::planning_validator::CollisionChecker" base_class_type="autoware::planning_validator::PluginInterface"/>
-</library>
