+++ planning/planning_validator/autoware_planning_validator_intersection_collision_checker/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_planning_validator_collision_checker</name>
+  <name>autoware_planning_validator_intersection_collision_checker</name>
   <version>1.0.0</version>
-  <description>The autoware_planning_validator_collision_checker package</description>
+  <description>The autoware_planning_validator_intersection_collision_checker package</description>
 
   <maintainer email="satoshi.ota@tier4.jp">Satoshi Ota</maintainer>
   <maintainer email="alqudah.mohammad@tier4.jp">Alqudah Mohammad</maintainer>
