+++ planning/planning_validator/autoware_planning_validator_intersection_collision_checker/plugins.xml
@@ -1,0 +1,3 @@
+<library path="autoware_planning_validator_intersection_collision_checker">
+  <class type="autoware::planning_validator::IntersectionCollisionChecker" base_class_type="autoware::planning_validator::PluginInterface"/>
+</library>
