+++ planning/planning_validator/autoware_planning_validator_intersection_collision_checker/src/utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/planning_validator_collision_checker/utils.hpp"
+#include "autoware/planning_validator_intersection_collision_checker/utils.hpp"
 
 #include <autoware/motion_utils/trajectory/trajectory.hpp>
 #include <autoware_lanelet2_extension/utility/utilities.hpp>
